Changed 32-bit variables of angles and reflectance models to 64-bit

+++ include/libbsdf/Brdf/Brdf.h
@@ -1,5 +1,5 @@
 // =================================================================== //
-// Copyright (C) 2014-2020 Kimura Ryo                                  //
+// Copyright (C) 2014-2023 Kimura Ryo                                  //
 //                                                                     //
 // This Source Code Form is subject to the terms of the Mozilla Public //
 // License, v. 2.0. If a copy of the MPL was not distributed with this //
@@ -76,44 +76,44 @@ public:
      * Computes incoming and outgoing directions of a Cartesian coordinate system
      * using a set of angle indices.
      */
-    virtual void getInOutDirection(int      index0,
-                                   int      index1,
-                                   int      index2,
-                                   int      index3,
-                                   Vec3*    inDir,
-                                   Vec3*    outDir) const = 0;
+    virtual void getInOutDirection(int   index0,
+                                   int   index1,
+                                   int   index2,
+                                   int   index3,
+                                   Vec3* inDir,
+                                   Vec3* outDir) const = 0;
 
     /*!
      * Converts from four angles to incoming and outgoing directions and
      * assigns them to \a inDir and \a outDir.
      */
-    virtual void toXyz(float angle0,
-                       float angle1,
-                       float angle2,
-                       float angle3,
-                       Vec3* inDir,
-                       Vec3* outDir) const = 0;
+    virtual void toXyz(double angle0,
+                       double angle1,
+                       double angle2,
+                       double angle3,
+                       Vec3*  inDir,
+                       Vec3*  outDir) const = 0;
 
     /*!
      * Converts from incoming and outgoing directions to four angles and
      * assigns them to \a angle0, \a angle1, \a angle2, and \a angle3.
      */
-    virtual void fromXyz(const Vec3&    inDir,
-                         const Vec3&    outDir,
-                         float*         angle0,
-                         float*         angle1,
-                         float*         angle2,
-                         float*         angle3) const = 0;
+    virtual void fromXyz(const Vec3& inDir,
+                         const Vec3& outDir,
+                         double*     angle0,
+                         double*     angle1,
+                         double*     angle2,
+                         double*     angle3) const = 0;
 
     /*!
      * Converts from incoming and outgoing directions to three angles for an isotropic BRDF and
      * assigns them to \a angle0, \a angle2, and \a angle3.
      */
-    virtual void fromXyz(const Vec3&    inDir,
-                         const Vec3&    outDir,
-                         float*         angle0,
-                         float*         angle2,
-                         float*         angle3) const = 0;
+    virtual void fromXyz(const Vec3& inDir,
+                         const Vec3& outDir,
+                         double*     angle0,
+                         double*     angle2,
+                         double*     angle3) const = 0;
 
     virtual std::string getAngle0Name() const = 0; /*!< Gets the name of angle0. */
     virtual std::string getAngle1Name() const = 0; /*!< Gets the name of angle1. */
